refactoring :: StudentsDomainInterface Transfer 나눔
