+++ Projects/Domain/StudentsDomain/Interface/DTO/Response/StudentInfoResponseDTO.swift
@@ -17,13 +17,3 @@ public struct StudentInfoResponseDTO: Decodable {
         case department
     }
 }
-
-public extension StudentInfoResponseDTO {
-    func toDomain() -> StudentInfoEntity {
-        StudentInfoEntity(
-            studentName: studentName,
-            studentGcn: studentGcn,
-            department: department
-        )
-    }
-}
