+++ Projects/Domain/StudentsDomain/Interface/DTO/Transfer/StudentInfoTransfer.swift
@@ -1,0 +1,11 @@
+import Foundation
+
+public extension StudentInfoResponseDTO {
+    func toDomain() -> StudentInfoEntity {
+        StudentInfoEntity(
+            studentName: studentName,
+            studentGcn: studentGcn,
+            department: department
+        )
+    }
+}
